feat: address pr feedback
