+++ test/appium-orchestrator/secrets.example.json
@@ -1,0 +1,16 @@
+{
+  "ios": {
+    "bundleId": "foobar",
+    "wdaBundleId": "foobar",
+    "udid": "foobar",
+    "xcodeOrgId": "foobar"
+  },
+  "sauce": {
+    "user": "foobar",
+    "key": "foobar",
+    "hostname": "foobar",
+    "port": 1234,
+    "baseUrl": "foobar",
+    "storageFilename": "storage:filename=foobar"
+  }
+}
