+++ test/appium-orchestrator/token.example.json
@@ -1,0 +1,3 @@
+{
+  "accessToken": "foobar"
+}
